chore: Moves localization utils files to folder
