+++ packages/core/src/utils/localization/validateLocaleForHostname.ts
@@ -1,5 +1,5 @@
 import storeConfig from 'discovery.config.default.js'
-import type { LocalesSettings } from '../typings/locales'
+import type { LocalesSettings } from '../../typings/locales'
 
 function getLocalesSettings(): LocalesSettings {
   return storeConfig.localization as LocalesSettings
